+++ refresh/index.js
@@ -37,7 +37,6 @@ var getServerSDKAsync = sdkHelper.getServerSDKAsync
 module.exports = Generator.extend({
   constructor: function () {
     Generator.apply(this, arguments)
-
     // Allow the user to specify where the specification file is
     this.option('specfile', {
       desc: 'The location of the specification file.',
@@ -149,80 +148,51 @@ module.exports = Generator.extend({
         this.env.error(chalk.red('Property appName is missing from the specification'))
       }
 
-      // Service configuration
-      this.services = this.spec.services || {}
-
       // Bluemix configuration
-      this.bluemix = {
-        backendPlatform: 'SWIFT',
-        name: helpers.sanitizeAppName(this.projectName),
-        server: {
-          name: helpers.sanitizeAppName(this.projectName),
-          env: {}
-        }
-      }
-      if (Object.keys(this.services).length > 0) {
-        this.bluemix.server.services = []
+      // Ensure minimal default bluemix properties are set
+      this.bluemix = this.spec.bluemix || {}
+      this.bluemix.backendPlatform = this.bluemix.backendPlatform || 'SWIFT'
+      this.bluemix.name = this.bluemix.name || helpers.sanitizeAppName(this.projectName)
+      this.bluemix.server = this.bluemix.server || {}
+      this.bluemix.server.name = this.bluemix.server.name || this.bluemix.name
+      this.bluemix.server.env = this.bluemix.server.env || {}
 
-        // Ensure every service has a credentials object, plan, & label
-        // and add the services to bluemix.server.services
-        Object.keys(this.services).forEach(function (serviceType) {
-          if (!Array.isArray(this.services[serviceType])) {
-            this.env.error(chalk.red(`Property services.${serviceType} must be an array`))
-          }
-          this.services[serviceType].forEach(function (service, index) {
-            // TODO: Further checking that service name is valid?
-            if (!service.name) {
-              this.env.error(chalk.red(`Service name is missing in spec for services.${serviceType}[${index}]`))
-            }
-            if (!service.label) {
-              service.label = helpers.getBluemixServiceLabel(serviceType)
-            }
-            if (!service.plan) {
-              service.plan = helpers.getBluemixDefaultPlan(serviceType)
-            }
-            service.credentials = service.credentials || {}
-            this.bluemix.server.services.push(service.name)
-          }.bind(this))
-        }.bind(this))
+      function isServiceProperty (value) {
+        if (Array.isArray(value)) {
+          // All elements of the service array must have a serviceInfo property
+          return value.filter(element => element.serviceInfo).length === value.length
+        }
+        return !!value.serviceInfo
+      }
+      function getServiceProperties (bluemixObject) {
+        var servicePropNames = Object.keys(bluemixObject).filter(prop => isServiceProperty(bluemixObject[prop]))
+        var serviceProperties = servicePropNames.reduce((memo, servicePropName) => {
+          return Object.assign(memo, { [servicePropName]: bluemixObject[servicePropName] })
+        }, {})
+        return serviceProperties
       }
 
-      if (typeof (this.spec.bluemix) === 'object') {
-        if (typeof (this.spec.bluemix.name) === 'string') {
-          this.bluemix.server.name = this.spec.bluemix.name
-          this.bluemix.name = this.spec.bluemix.name
+      this.services = getServiceProperties(this.bluemix)
+      Object.keys(this.services).forEach(serviceType => {
+        var serviceOrServices = this.services[serviceType]
+        if (Array.isArray(serviceOrServices)) {
+          serviceOrServices.forEach((service, index) => {
+            var updatedService = helpers.sanitizeServiceAndFillInDefaults(serviceType, service)
+            helpers.validateServiceFields(serviceType, updatedService)
+            this.bluemix[serviceType][index] = updatedService
+            this.services[serviceType][index] = updatedService
+          })
+        } else {
+          var updatedService = helpers.sanitizeServiceAndFillInDefaults(serviceType, serviceOrServices)
+          helpers.validateServiceFields(serviceType, updatedService)
+          this.bluemix[serviceType] = updatedService
+          this.services[serviceType] = updatedService
         }
-        if (this.spec.bluemix.server) {
-          if (typeof (this.spec.bluemix.server.host) === 'string') {
-            this.bluemix.server.host = this.spec.bluemix.server.host
-          }
-          if (typeof (this.spec.bluemix.server.domain) === 'string') {
-            this.bluemix.server.domain = this.spec.bluemix.server.domain
-          }
-          if (typeof (this.spec.bluemix.server.memory) === 'string') {
-            this.bluemix.server.memory = this.spec.bluemix.server.memory
-          }
-          if (typeof (this.spec.bluemix.server.disk_quota) === 'string') {
-            this.bluemix.server.disk_quota = this.spec.bluemix.server.disk_quota
-          }
-          if (typeof (this.spec.bluemix.server.instances) === 'number') {
-            this.bluemix.server.instances = this.spec.bluemix.server.instances
-          }
-          if (typeof (this.spec.bluemix.server.namespace) === 'string') {
-            this.bluemix.server.namespace = this.spec.bluemix.server.namespace
-          }
-          if (typeof (this.spec.bluemix.server.cloudDeploymentType) === 'string') {
-            this.bluemix.server.cloudDeploymentType = this.spec.bluemix.server.cloudDeploymentType
-          }
-          if (typeof (this.spec.bluemix.server.cloudDeploymentOptions) === 'object') {
-            this.bluemix.server.cloudDeploymentOptions = this.spec.bluemix.server.cloudDeploymentOptions
-          }
-        }
-        if (typeof (this.spec.bluemix.openApiServers) === 'object') {
-          this.openApiServers = this.spec.bluemix.openApiServers
-        }
-      }
+      })
 
+      if (typeof (this.bluemix.openApiServers) === 'object') {
+        this.openApiServers = this.bluemix.openApiServers
+      }
       // Docker configuration
       this.docker = (this.spec.docker === true)
 
@@ -260,7 +230,7 @@ module.exports = Generator.extend({
       this.metrics = (this.spec.metrics === true || undefined)
 
       // Autoscaling implies monitoring
-      if (this.services.autoscaling && this.services.autoscaling.length > 0) {
+      if (this.bluemix.autoscaling) {
         this.metrics = true
       }
 
@@ -300,23 +270,32 @@ module.exports = Generator.extend({
       this.itemsToIgnore = []
 
       // Package dependencies to add to Package.swift
-      // eg this.dependencies.push('.Package(url: "https://github.com/IBM-Swift/Kitura.git", majorVersion: 1, minor: 7),')
+      // eg this.dependencies.push('.package(url: "https://github.com/IBM-Swift/Kitura.git", .upToNextMinor(from : "1.7.0")),')
       this.dependencies = []
+
+      // Module Dependencies to add to Package.swift
+      this.modules = []
 
       // Initialization code to add to Application.swift by code block
       // eg this.appInitCode.services.push('try initializeServiceCloudant()')
       this.appInitCode = {
         capabilities: [],
         services: [],
+        service_imports: [],
+        service_variables: [],
         middlewares: [],
         endpoints: []
       }
 
       if (this.web) this.appInitCode.middlewares.push('router.all(middleware: StaticFileServer())')
-      if (this.appType === 'crud') this.appInitCode.endpoints.push('try initializeCRUDResources(cloudEnv: cloudEnv, router: router)')
+      if (this.appType === 'crud') {
+        this.appInitCode.endpoints.push('try initializeCRUDResources(cloudEnv: cloudEnv, router: router)')
+        this.dependencies.push('.package(url: "https://github.com/IBM-Swift/SwiftyJSON.git", from: "17.0.0"),')
+      }
       if (this.metrics) {
-        this.appInitCode.capabilities.push('initializeMetrics()')
-        this.dependencies.push('.Package(url: "https://github.com/RuntimeTools/SwiftMetrics.git", majorVersion: 1),')
+        this.modules.push('"SwiftMetrics"')
+        this.appInitCode.capabilities.push('initializeMetrics(app: self)')
+        this.dependencies.push('.package(url: "https://github.com/RuntimeTools/SwiftMetrics.git", from: "2.0.0"),')
       }
     },
 
@@ -416,64 +395,22 @@ module.exports = Generator.extend({
 
   configuring: function () {
     if (this.existingProject) return
-
-    var bluemixOption = {
-      backendPlatform: 'SWIFT',
-      name: this.projectName, // TODO: check this is the right name
-      server: {
-        name: this.projectName // TODO: check this is the right name
-      }
-    }
-    // NOTE(tunniclm): Set the domain based on the push notification
-    // region to expose it to the service enablement subgenerator
-    if (this.services.pushnotifications && this.services.pushnotifications.length > 0) {
-      var push = this.services.pushnotifications[0]
-      switch (push.region) {
-        case 'UK': bluemixOption.server.domain = 'eu-gb.bluemix.net'; break
-        case 'SYDNEY': bluemixOption.server.domain = 'au-syd.bluemix.net'; break
-        case 'US_SOUTH': bluemixOption.server.domain = 'ng.bluemix.net'; break
-        // default: don't alter domain
-      }
-    }
-    // NOTE(tunniclm): Convert our format for specifying services
-    // into the one used by generator-ibm-service-enablement
-    var serviceMapping = {
-      'appid': 'auth',
-      'objectstorage': 'objectStorage',
-      'cloudant': 'cloudant',
-      'watsonconversation': 'conversation',
-      'redis': 'redis',
-      'mongodb': 'mongodb',
-      'postgresql': 'postgresql',
-      'alertnotification': 'alertnotification',
-      'pushnotifications': 'push',
-      'autoscaling': 'autoscaling'
-    }
-    Object.keys(serviceMapping).forEach(serviceType => {
-      var bluemixServiceProperty = serviceMapping[serviceType]
-      var servicesOfType = this.services[serviceType]
-      if (servicesOfType && servicesOfType.length > 0) {
-        // NOTE: for now only handle 1 service
-        var service = helpers.sanitizeServiceAndFillInDefaults(serviceType, servicesOfType[0])
-        bluemixOption[bluemixServiceProperty] = service.credentials || {}
-        bluemixOption[bluemixServiceProperty].serviceInfo = {
-          name: servicesOfType[0].name,
-          label: servicesOfType[0].label,
-          plan: servicesOfType[0].plan
-        }
-      }
-    })
     this.composeWith(require.resolve('generator-ibm-service-enablement'), {
       quiet: true,
-      bluemix: JSON.stringify(bluemixOption),
+      bluemix: JSON.stringify(this.bluemix),
       parentContext: {
         injectIntoApplication: options => {
           if (options.capability) this.appInitCode.capabilities.push(options.capability)
           if (options.service) this.appInitCode.services.push(options.service)
+          if (options.service_import) this.appInitCode.service_imports.push(options.service_import)
+          if (options.service_variable) this.appInitCode.service_variables.push(options.service_variable)
           if (options.endpoint) this.appInitCode.endpoints.push(options.endpoint)
           if (options.middleware) this.appInitCode.middlewares.push(options.middleware)
         },
-        injectDependency: dependency => { this.dependencies.push(dependency) }
+        injectDependency: dependency => { this.dependencies.push(dependency) },
+        injectModules: modules => {
+          this.modules.push(modules)
+        }
       }
     })
   },
@@ -834,14 +771,18 @@ module.exports = Generator.extend({
       endpointNames = endpointNames.concat(resourceNames)
     }
     if (this.healthcheck) {
+      this.modules.push('"Health"')
       endpointNames.push('Health')
-      this.dependencies.push('.Package(url: "https://github.com/IBM-Swift/Health.git", majorVersion: 0),')
+      this.dependencies.push('.package(url: "https://github.com/IBM-Swift/Health.git", from: "0.0.0"),')
     }
 
-    var initCodeForEndpoints = endpointNames.map(name => `initialize${name}Routes()`)
+    var initCodeForEndpoints = endpointNames.map(name => `initialize${name}Routes(app: self)`)
     this.appInitCode.endpoints = this.appInitCode.endpoints.concat(initCodeForEndpoints)
 
-    if (this.hostSwagger) this.appInitCode.endpoints.push(`initializeSwaggerRoutes(path: projectPath + "/definitions/${this.projectName}.yaml")`)
+    if (this.hostSwagger) {
+      this.appInitCode.endpoints.push(`initializeSwaggerRoutes(app: self)`)
+      this.swaggerPath = `let swaggerPath = projectPath + "/definitions/${this.projectName}.yaml"`
+    }
   },
 
   generateSDKs: function () {
@@ -849,7 +790,6 @@ module.exports = Generator.extend({
     var shouldGenerateClient = (!!this.openApiDocumentBytes)
     var shouldGenerateServer = (this.serverSwaggerFiles.length > 0)
     if (!shouldGenerateClientWithModel && !shouldGenerateClient && !shouldGenerateServer) return
-
     this.log(chalk.green('Generating SDK(s) from swagger file(s)...'))
     var generationTasks = []
     if (shouldGenerateClientWithModel) generationTasks.push(generateClientAsync.call(this, this.swagger))
@@ -955,8 +895,8 @@ module.exports = Generator.extend({
           {
             appType: this.appType,
             generatedModule: this.generatedModule,
-            bluemix: this.bluemix,
             appInitCode: this.appInitCode,
+            swaggerPath: this.swaggerPath,
             web: this.web,
             healthcheck: this.healthcheck,
             basepath: this.parsedSwagger && this.parsedSwagger.basepath
@@ -1046,20 +986,19 @@ module.exports = Generator.extend({
             appName: this.projectName,
             executableName: this.executableModule,
             generatorVersion: this.generatorVersion,
-            bluemix: this.bluemix,
             web: this.web,
             docker: this.docker,
             hostSwagger: this.hostSwagger,
             exampleEndpoints: this.exampleEndpoints,
             metrics: this.metrics,
-            autoscaling: this.services.autoscaling && this.services.autoscaling.length > 0,
-            cloudant: this.services.cloudant && this.services.cloudant.length > 0,
-            redis: this.services.redis && this.services.redis.length > 0,
-            objectstorage: this.services.objectstorage && this.services.objectstorage.length > 0,
-            appid: this.services.appid && this.services.appid.length > 0,
-            watsonconversation: this.services.watsonconversation && this.services.watsonconversation.length > 0,
-            alertnotification: this.services.alertnotification && this.services.alertnotification.length > 0,
-            pushnotifications: this.services.pushnotifications && this.services.pushnotifications.length > 0
+            autoscaling: !!this.bluemix.autoscaling,
+            cloudant: this.bluemix.cloudant && this.bluemix.cloudant.length > 0,
+            redis: !!this.bluemix.redis,
+            objectStorage: this.bluemix.objectStorage && this.bluemix.objectStorage.length > 0,
+            auth: !!this.bluemix.auth,
+            conversation: !!this.bluemix.conversation,
+            alertNotification: !!this.bluemix.alertNotification,
+            push: !!this.bluemix.push
           }
         )
         this.fs.write(this.destinationPath('Sources', this.applicationModule, 'Routes', '.keep'), '')
@@ -1090,13 +1029,11 @@ module.exports = Generator.extend({
     createCRUD: function () {
       if (this.appType !== 'crud') return
 
-      if (this.bluemix) {
-        if (!this.fs.exists(this.destinationPath('README.md'))) {
-          this.fs.copy(
-            this.templatePath('bluemix', 'README.md'),
-            this.destinationPath('README.md')
-          )
-        }
+      if (!this.fs.exists(this.destinationPath('README.md'))) {
+        this.fs.copy(
+          this.templatePath('bluemix', 'README.md'),
+          this.destinationPath('README.md')
+        )
       }
 
       // Add the models to the spec
@@ -1117,8 +1054,10 @@ module.exports = Generator.extend({
         var serviceDef = null
         Object.keys(services).forEach(function (serviceType) {
           if (serviceDef) return
-          services[serviceType].forEach(function (service) {
-            if (service.name && (service.name === serviceName)) {
+          var serviceOrServices = services[serviceType]
+          if (!Array.isArray(serviceOrServices)) serviceOrServices = [serviceOrServices]
+          serviceOrServices.forEach(function (service) {
+            if (service.serviceInfo.name && (service.serviceInfo.name === serviceName)) {
               serviceDef = {
                 service: service,
                 type: serviceType
@@ -1144,7 +1083,7 @@ module.exports = Generator.extend({
       this.fs.copyTpl(
         this.templatePath('crud', 'AdapterFactory.swift'),
         this.destinationPath('Sources', this.generatedModule, 'AdapterFactory.swift'),
-        { models: this.models, crudService: crudService, bluemix: this.bluemix }
+        { models: this.models, crudService: crudService }
       )
       this.fs.copy(
         this.templatePath('crud', 'AdapterError.swift'),
@@ -1275,13 +1214,19 @@ module.exports = Generator.extend({
     },
 
     writeDockerFiles: function () {
-      if (!this.docker || this.existingProject || !this.bluemix) return
+      if (!this.docker || this.existingProject) return
       this.composeWith(require.resolve('generator-ibm-cloud-enablement/generators/dockertools'), { force: this.force, bluemix: this.bluemix })
     },
 
     writeBluemixDeploymentFiles: function () {
       if (this.existingProject) return
-      this.bluemix.services = this.services
+      // Go through the bluemix object and pull out the serviceInfo for each of
+      // the services to pass down to cloud enablement
+      this.bluemix.services = {}
+      Object.keys(this.services).forEach(prop => {
+        var services = Array.isArray(this.services[prop]) ? this.services[prop] : [this.services[prop]]
+        this.bluemix.services[prop] = services.map(service => service.serviceInfo)
+      })
       this.composeWith(require.resolve('generator-ibm-cloud-enablement/generators/deployment'), { force: this.force, bluemix: this.bluemix, repoType: this.repoType })
     },
 
@@ -1302,7 +1247,8 @@ module.exports = Generator.extend({
             generatedModule: this.generatedModule,
             applicationModule: this.applicationModule,
             sdkTargets: this.sdkTargets,
-            dependencies: this.dependencies
+            dependencies: this.dependencies,
+            modules: this.modules
           }
         )
       })